chore(process::resources): change Ri to u32 instead of usize

+++ crates/kernel/src/process/resources.rs
@@ -14,7 +14,7 @@ use hashbrown::HashMap;
 use safa_abi::errors::ErrorStatus;
 
 /// A resource ID
-pub type Ri = usize;
+pub type Ri = u32;
 
 pub trait Resource: Any {
     /// Performs a write operation on the resource.
